# api/@typespec/openapi3/openapi.yaml
openapi: 3.0.0
info:
  title: Calendar Booking API
  version: 0.0.0
tags: []
paths:
  /admin/bookings:
    get:
      operationId: Bookings_listUpcoming
      description: List upcoming bookings across all event types.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
  /admin/event-types:
    post:
      operationId: EventTypes_create
      description: Create a new event type for the predefined owner profile. slotStepMinutes must be 15 or 30.
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                durationMinutes:
                  type: integer
                  format: int32
                slotStepMinutes:
                  type: integer
                  format: int32
              required:
                - name
                - description
                - durationMinutes
                - slotStepMinutes
    get:
      operationId: EventTypes_list
      description: List all event types of the predefined owner profile.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventType'
  /admin/event-types/{id}:
    patch:
      operationId: EventTypeItem_update
      description: Update an existing event type. durationMinutes and slotStepMinutes are editable. slotStepMinutes must be 15 or 30.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventType'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                durationMinutes:
                  type: integer
                  format: int32
                slotStepMinutes:
                  type: integer
                  format: int32
  /bookings:
    post:
      operationId: PublicBookings_create
      description: Create a booking for a guest without registration. Two bookings cannot share the same start/end time, even across different event types.
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: The request conflicts with the current state of the server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                eventTypeId:
                  type: string
                startAt:
                  type: string
                  format: date-time
                guestName:
                  type: string
                guestEmail:
                  type: string
                guestPhone:
                  type: string
                guestComment:
                  type: string
              required:
                - eventTypeId
                - startAt
                - guestName
                - guestEmail
  /event-types:
    get:
      operationId: PublicEventTypes_list
      description: List all event types available for guests.
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventType'
  /event-types/{id}/slots:
    get:
      operationId: PublicSlots_list
      description: List available slots for the next 14 days from now. Optional from/to query can narrow this window but must remain within 14 days.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
          explode: false
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Slot'
        '400':
          description: The server could not understand the request due to invalid syntax.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: The server cannot find the requested resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
    Booking:
      type: object
      required:
        - id
        - eventTypeId
        - startAt
        - endAt
        - guestName
        - guestEmail
        - createdAt
      properties:
        id:
          type: string
        eventTypeId:
          type: string
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        guestName:
          type: string
        guestEmail:
          type: string
        guestPhone:
          type: string
        guestComment:
          type: string
        createdAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        message:
          type: string
    EventType:
      type: object
      required:
        - id
        - name
        - description
        - durationMinutes
        - slotStepMinutes
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
          format: int32
        slotStepMinutes:
          type: integer
          format: int32
    Slot:
      type: object
      required:
        - startAt
        - endAt
        - isAvailable
      properties:
        startAt:
          type: string
          format: date-time
        endAt:
          type: string
          format: date-time
        isAvailable:
          type: boolean
servers:
  - url: http://localhost:3000
    description: Local development server
    variables: {}